Use Pyglet with FFmpeg on all platforms.
The current idea is that arcade would provide the necessary binaries
for FFmpeg on all platforms. It's the case here for Windows and MacOS.
I'm not sure how things work on Linux. If we need to rely on the OS
to have those binaries, then we'll need additional checks to figure
out if the correct version is available. If it's not available, what
do we want to do? Still provide some default sound with the functions
` _load_sound_unix` and `_play_sound_unix`?